1)添加了气象云图下载功能的控制选项，下载可以选择工作站可以不下载气象云图了； 2）客户端CPU和内存占用率显示进度条； 3）升级项目文件到VS2017

+++ CC-Client/CC-Client/recvfilethread.cpp
@@ -43,7 +43,7 @@ void RecvFileThread::run()
 #endif
 				if (pf != NULL)
 				{
-					size_t size = s->fileProxy->getSize(srcFile.toStdWString());
+					qint64 size = s->fileProxy->getSize(srcFile.toStdWString());
 					emit notifyFileSize(s, size);
 					getData(s, 0, pf);					
 				}
@@ -74,7 +74,7 @@ void RecvFileThread::run()
 }
 
 //获取数据
-void RecvFileThread::getData(StationInfo* s, size_t position, FILE* pf)
+void RecvFileThread::getData(StationInfo* s, qint64 position, FILE* pf)
 {
 	std::vector< Ice::Byte > inParams;
